+++ chalicelib/processors/ComplianceProcessor.py
@@ -1,5 +1,6 @@
 from chalicelib.processors.GenericProcessor import GenericProcessor
 from chalicelib.version import API_VERSION
+from chalicelib.usage import OpenAIDefaults
 
 
 class ComplianceProcessor(GenericProcessor):
@@ -7,7 +8,8 @@ class ComplianceProcessor(GenericProcessor):
         super().__init__(API_VERSION, {
             'main': 'compliance.prompt',
             'role_system': 'compliance-role-system.prompt'
-        })
+        }, {'model': OpenAIDefaults.boost_default_gpt_model,
+            'temperature': OpenAIDefaults.temperature_medium_with_explanation})
 
     def get_chunkable_input(self) -> str:
         return 'code'